requirements.txt
# C-to-Rust migration framework
# Install: pip install -r requirements.txt
#
# GPU note: the default torch below is CPU-only.
# For CUDA training install the matching wheel first, e.g.:
#   pip install torch --index-url https://download.pytorch.org/whl/cu121

# --- OpenAI engine (C2RustAI.py) ---
openai>=1.0.0
python-dotenv>=1.0.0

# --- Static analysis (analyzer.py) ---
libclang>=16.0.0

# --- Local model engine (C2RustLocal.py) ---
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.28.0      # required by transformers for device_map="auto"
peft>=0.10.0            # LoRA adapters
trl>=0.12.0             # GRPOTrainer (processing_class param added in 0.12)
datasets>=2.18.0        # training data loader
bitsandbytes>=0.41.0    # QLoRA 4-bit quantization (requires CUDA)
